HR ANALYTICS DASHBOARD
Incorporating a blend of intuitive visualizations, the HR Analytics Dashboard serves as a powerful tool for HR professionals aspiring to make data-driven decisions. Whether it's understanding attrition trends over time, by age, or in relation to salary, this dashboard empowers users to navigate the complexities of workforce dynamics with precision.